+++ scripts/reinforcement_learning/rl_games/train.py
@@ -64,6 +64,7 @@ import math
 import os
 import random
 from datetime import datetime
+from gymnasium.wrappers import RecordVideo
 
 import omni
 from rl_games.common import env_configurations, vecenv
@@ -80,12 +81,15 @@ from isaaclab.envs import (
 from isaaclab.utils.assets import retrieve_file_path
 from isaaclab.utils.dict import print_dict
 from isaaclab.utils.io import dump_pickle, dump_yaml
-from isaaclab.utils.recorder import RecordVideo
+from isaaclab.utils.recorder import stop_recording
 
 from isaaclab_rl.rl_games import RlGamesGpuEnv, RlGamesVecEnvWrapper
 
 import isaaclab_tasks  # noqa: F401
 from isaaclab_tasks.utils.hydra import hydra_task_config
+
+# monkey-patch the stop_recording function to fix memory leak error (see https://github.com/Farama-Foundation/Gymnasium/pull/1444)
+RecordVideo.stop_recording = stop_recording
 
 # PLACEHOLDER: Extension template (do not remove this comment)